Restructure component so that accordion works
